decompose core
