+++ backend/lessons/scorm/engine/core/datasets.py
@@ -16,16 +16,31 @@ class DataSetCore:
             raise ElementTypeError(
                 'Ожидался Element',
             )
-        self.element = element
+        self._element = element
         self.prefix = str(prefix)
+
+    @property
+    def element(self):
+        return self._element
 
     def __getitem__(self, index: str) -> list[ET.Element]:
         index_p = self.prefix + str(index)
-        element = self.element.find(index_p)
+        element = self._element.find(index_p)
         if element is None:
-            element = self.element.findall(index_p)
+            element = self._element.findall(index_p)
+            if element:
+                element = [DataSetCore(
+                    element=el,
+                    prefix=self.prefix,
+                    )
+                           for el
+                           in element]
         else:
-            element = [element]
+            element = [DataSetCore(
+                element=element,
+                prefix=self.prefix,
+            )
+            ]
         return element
 
     def __repr__(self):
